feat(extension): lw-6915 refactor send tx form
