+++ apps/browser-extension-wallet/src/views/browser-view/features/send-transaction/components/Form/Form.module.scss
@@ -1,16 +1,6 @@
 @import '../../../../../../../../../packages/common/src/ui/styles/theme.scss';
 @import '../../../../../../../../../packages/common/src/ui/styles/abstracts/typography';
 
-.outputRowContainer {
-  display: flex;
-  flex-direction: column;
-  gap: size_unit(2);
-  padding-bottom: size_unit(2.5);
-  @media (max-width: $breakpoint-popup) {
-    padding-bottom: size_unit(2);
-  }
-}
-
 .costsContainer {
   display: flex;
   flex-direction: column;
@@ -20,6 +10,7 @@
 
   .titleContainer {
     color: var(--text-color-primary);
+
     .title {
       @include text-subHeading-bold;
       margin: 0;
@@ -65,10 +56,12 @@
     &:hover {
       background: var(--light-mode-light-grey);
     }
+
     &:disabled {
       background-color: var(--color-white, #ffffff);
       opacity: 0.3;
     }
+
     .bundleIcon {
       font-size: var(--body, 16px);
     }
